+++ CHANGELOG.md
@@ -9,6 +9,58 @@ Internal builds may append content to the Unreleased section.
 Only write entries that are worth mentioning to users.
 -->
 
+## [0.59] - 2025-11-28
+
+- Core: Move context file location to `.kimi/sessions/{workdir_md5}/{session_id}/context.jsonl`
+- Lib: Move `WireMessage` type alias to `kimi_cli.wire.message`
+- Lib: Add `kimi_cli.wire.message.Request` type alias request messages (which currently only includes `ApprovalRequest`)
+- Lib: Add `kimi_cli.wire.message.is_event`, `is_request` and `is_wire_message` utility functions to check the type of wire messages
+- Lib: Add `kimi_cli.wire.serde` module for serialization and deserialization of wire messages
+- Lib: Change `StatusUpdate` Wire message to not using `kimi_cli.soul.StatusSnapshot`
+- Core: Record Wire messages to a JSONL file in session directory
+- Core: Introduce `TurnBegin` Wire message to mark the beginning of each agent turn
+- UI: Print user input again with a panel in shell mode
+- Lib: Add `Session.dir` property to get the session directory path
+- UI: Improve "Approve for session" experience when there are multiple parallel subagents
+- Wire: Reimplement Wire server mode (which is enabled with `--wire` option)
+- Lib: Rename `ShellApp` to `Shell`, `PrintApp` to `Print`, `ACPServer` to `ACP` and `WireServer` to `WireOverStdio` for better consistency
+- Lib: Rename `KimiCLI.run_shell_mode` to `run_shell`, `run_print_mode` to `run_print`, `run_acp_server` to `run_acp`, and `run_wire_server` to `run_wire_stdio` for better consistency
+- Lib: Add `KimiCLI.run` method to run a turn with given user input and yield Wire messages
+- Print: Fix stream-json print mode not flushing output properly
+- LLM: Improve compatibility with some OpenAI and Anthropic API providers
+- Core: Fix chat provider error after compaction when using Anthropic API
+
+## [0.58] - 2025-11-21
+
+- Core: Fix field inheritance of agent spec files when using `extend`
+- Core: Support using MCP tools in subagents
+- Tool: Add `CreateSubagent` tool to create subagents dynamically (not enabled in default agent)
+- Tool: Use MoonshotFetch service in `FetchURL` tool for Kimi for Coding plan
+- Tool: Truncate Grep tool output to avoid exceeding token limit
+
+## [0.57] - 2025-11-20
+
+- LLM: Fix Google GenAI provider when thinking toggle is not on
+- UI: Improve approval request wordings
+- Tool: Remove `PatchFile` tool
+- Tool: Rename `Bash`/`CMD` tool to `Shell` tool
+- Tool: Move `Task` tool to `kimi_cli.tools.multiagent` module
+
+## [0.56] - 2025-11-19
+
+- LLM: Add support for Google GenAI provider
+
+## [0.55] - 2025-11-18
+
+- Lib: Add `kimi_cli.app.enable_logging` function to enable logging when directly using `KimiCLI` class
+- Core: Fix relative path resolution in agent spec files
+- Core: Prevent from panic when LLM API connection failed
+- Tool: Optimize `FetchURL` tool for better content extraction
+- Tool: Increase MCP tool call timeout to 60 seconds
+- Tool: Provide better error message in `Glob` tool when pattern is `**`
+- ACP: Fix thinking content not displayed properly
+- UI: Minor UI improvements in shell mode
+
 ## [0.54] - 2025-11-13
 
 - Lib: Move `WireMessage` from `kimi_cli.wire.message` to `kimi_cli.wire`